# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Anti-Scam Trainer API
  version: 1.0.0
  description: API интерактивного тренажёра безопасных сделок.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /v1/scenarios:
    get:
      operationId: listScenarios
      responses:
        "200":
          description: Сценарии и прогресс в их текущих версиях
          content:
            application/json:
              schema:
                type: object
                required: [scenarios]
                properties:
                  scenarios:
                    type: array
                    items: { $ref: "#/components/schemas/ScenarioSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /v1/users/me:
    get:
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
...
    Unauthorized:
      description: Отсутствует или недействителен Bearer JWT
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    InternalError:
      description: Внутренняя ошибка сервера
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Score:
      type: object
      required: [points, maxPoints, percent]
      properties:
        points: { type: integer, minimum: 0 }
        maxPoints: { type: integer, minimum: 1 }
        percent: { type: integer, minimum: 0, maximum: 100 }
    ScenarioVersion:
      type: object
      required: [number, passPercent, publishedAt]
      properties:
        number: { type: integer, minimum: 1 }
        passPercent: { type: integer, minimum: 1, maximum: 100 }
        publishedAt: { type: string, format: date-time }
    CurrentVersionProgress:
      type: object
      required: [completed, passed, attemptsCount, bestScore, activeAttemptId]
      properties:
        completed: { type: boolean }
        passed: { type: boolean }
        attemptsCount: { type: integer, minimum: 0 }
        bestScore: { oneOf: [ { $ref: "#/components/schemas/Score" }, { type: "null" } ] }
        activeAttemptId: { type: [string, "null"], format: uuid }
    ScenarioSummary:
      type: object
      required: [slug, title, role, riskModule, currentVersion, currentProgress]
      properties:
        slug: { type: string }
        title: { type: string }
        role: { type: string, enum: [buyer, seller] }
        riskModule: { type: string }
        currentVersion: { $ref: "#/components/schemas/ScenarioVersion" }
        currentProgress: { $ref: "#/components/schemas/CurrentVersionProgress" }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
    User:
      type: object
      required:
        - id
        - nickname
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time